+++ build.gradle
@@ -1,6 +1,7 @@
 plugins {
     id 'java'
     id 'application'
+    id "com.diffplug.spotless" version "5.9.0"
 }
 
 group 'org.pwl'
@@ -21,6 +22,7 @@ test {
         logger.lifecycle("Gradle running test: ${descriptor}")
     }
 }
+
 dependencies {
     //testCompile group: 'junit', name: 'junit', version: '4.12'
     testCompile group: 'org.testng', name: 'testng', version: '7.3.0'
@@ -30,4 +32,17 @@ dependencies {
 
 
 }
+
+spotless {
+    format 'misc', {
+        target '*.gradle', '*.md', '.gitignore'
+        trimTrailingWhitespace()
+        indentWithSpaces()
+        endWithNewline()
+    }
+    java {
+        googleJavaFormat('1.8').aosp()
+    }
+}
+
 applicationDefaultJvmArgs = ["-Dtestng.dtd.http=true"]
