+++ src/test/java/org/pwl/framework/OverviewTabTests.java
@@ -1,0 +1,15 @@
+package org.pwl.framework;
+
+import static org.testng.Assert.*;
+
+import org.openqa.selenium.By;
+import org.testng.annotations.*;
+
+public class OverviewTabTests extends BaseTestClass {
+    @Test
+    void usernameIsCorrectOnOverviewTab() {
+        driver.get(BASE_URL + USER);
+        String actualUsername = driver.findElement(By.className("p-nickname")).getText();
+        assertEquals(USER, actualUsername);
+    }
+}
